Implementa feats da SP2 e aprimora feats da SP1

+++ src/main/java/com/recrutaibackend/dto/AddressResponse.java
@@ -1,7 +1,7 @@
 package com.recrutaibackend.dto;
 
 public record AddressResponse(
-        Integer id,
+        long id,
         String streetAddress,
         String city,
         String state,
