+++ src/main/java/com/recrutaibackend/repository/EmailVerificationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EmailVerificationRepository extends JpaRepository<EmailVerification, Integer> {
+public interface EmailVerificationRepository extends JpaRepository<EmailVerification, Long> {
 
     Optional<EmailVerification> findByCode(String code);
