+++ src/main/java/com/recrutaibackend/repository/InterviewRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InterviewRepository extends JpaRepository<Interview, Integer> {
+public interface InterviewRepository extends JpaRepository<Interview, Long> {
 }
