+++ src/main/java/com/recrutaibackend/repository/VacancyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface VacancyRepository extends JpaRepository<Vacancy, Integer> {
+public interface VacancyRepository extends JpaRepository<Vacancy, Long> {
 
     List<Vacancy> findAllByTitleContainsIgnoreCase(String title);
 }
